requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
bitsandbytes==0.48.2
certifi==2025.11.12
charset-normalizer==3.4.4
cut-cross-entropy==25.1.1
datasets==4.3.0
deepseek-vl==0.0.1.dev0
diffusers==0.35.2
dill==0.4.0
docstring_parser==0.17.0
editdistance==0.8.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
gdown==5.2.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.6
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
peft==0.18.0
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
sentencepiece==0.2.1
setuptools==80.9.0
shtab==1.8.0
six==1.17.0
soupsieve==2.8
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tokenizers==0.22.1
torch==2.9.1
torchao==0.14.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.2
triton==3.5.1
trl==0.24.0
typeguard==4.4.4
typing_extensions==4.15.0
tyro==0.9.35
tzdata==2025.2
unsloth==2025.11.6
unsloth_zoo==2025.12.1
urllib3==2.6.0
Werkzeug==3.1.4
wheel==0.45.1
xformers==0.0.33.post2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
